requirements.txt
aiosqlite==0.20.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
bcrypt==4.0.1
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.115.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.4
loguru==0.7.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
orjson==3.11.7
passlib==1.7.4
pyasn1==0.6.2
pydantic==2.9.2
pydantic-extra-types==2.11.0
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.19.2
PyJWT==2.11.0
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.22
PyYAML==6.0.3
rich==14.3.2
rich-toolkit==0.18.1
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.51.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.35
starlette==0.38.6
typer==0.21.1
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.31.0
watchfiles==1.1.1
websockets==16.0
win32_setctime==1.2.0
